+++ platform/darwin/src/MGLPolyline.mm
@@ -1,6 +1,7 @@
 #import "MGLPolyline.h"
 
 #import "MGLMultiPoint_Private.h"
+#import "MGLGeometry_Private.h"
 
 @implementation MGLPolyline
 
@@ -23,6 +24,42 @@
     shapeProperties.set<mbgl::LineAnnotationProperties>(lineProperties);
     
     return shapeProperties;
+}
+
+@end
+
+@interface MGLMultiPolyline ()
+
+@property (nonatomic, copy, readwrite) NS_ARRAY_OF(MGLPolyline *) *polylines;
+
+@end
+
+@implementation MGLMultiPolyline {
+    MGLCoordinateBounds _overlayBounds;
+}
+
+@synthesize overlayBounds = _overlayBounds;
+
++ (instancetype)multiPolylineWithPolylines:(NS_ARRAY_OF(MGLPolyline *) *)polylines {
+    return [[self alloc] initWithPolylines:polylines];
+}
+
+- (instancetype)initWithPolylines:(NS_ARRAY_OF(MGLPolyline *) *)polylines {
+    if (self = [super init]) {
+        _polylines = polylines;
+        
+        mbgl::LatLngBounds bounds = mbgl::LatLngBounds::empty();
+        
+        for (MGLPolyline *polyline in _polylines) {
+            bounds.extend(MGLLatLngBoundsFromCoordinateBounds(polyline.overlayBounds));
+        }
+        _overlayBounds = MGLCoordinateBoundsFromLatLngBounds(bounds);
+    }
+    return self;
+}
+
+- (BOOL)intersectsOverlayBounds:(MGLCoordinateBounds)overlayBounds {
+    return MGLLatLngBoundsFromCoordinateBounds(_overlayBounds).intersects(MGLLatLngBoundsFromCoordinateBounds(overlayBounds));
 }
 
 @end