+++ platform/darwin/src/MGLShapeCollection.m
@@ -1,0 +1,21 @@
+#import "MGLShapeCollection.h"
+
+@implementation MGLShapeCollection
+
++ (instancetype)shapeCollectionWithShapes:(NS_ARRAY_OF(MGLShape <MGLAnnotation> *) *)shapes {
+    return [[self alloc] initWithShapes:shapes];
+}
+
+- (instancetype)initWithShapes:(NS_ARRAY_OF(MGLShape <MGLAnnotation> *) *)shapes {
+    if (self = [super init]) {
+        NSAssert(shapes.count, @"Cannot create an empty shape collection");
+        _shapes = shapes.copy;
+    }
+    return self;
+}
+
+- (CLLocationCoordinate2D)coordinate {
+    return _shapes.firstObject.coordinate;
+}
+
+@end