+++ platform/osx/app/MapDocument.m
@@ -420,7 +420,8 @@ static const CLLocationCoordinate2D WorldTourDestinations[] = {
         if (!NSPointInRect([gestureRecognizer locationInView:self.mapView.compass], self.mapView.compass.bounds)
             && !NSPointInRect([gestureRecognizer locationInView:self.mapView.zoomControls], self.mapView.zoomControls.bounds)
             && !NSPointInRect([gestureRecognizer locationInView:self.mapView.attributionView], self.mapView.attributionView.bounds)) {
-            [self dropPinAtPoint:location];
+            NSArray *features = [self.mapView visibleFeaturesAtPoint:location];
+            [self.mapView addAnnotations:features];
         }
     }
 }
@@ -436,9 +437,17 @@ static const CLLocationCoordinate2D WorldTourDestinations[] = {
 }
 
 - (DroppedPinAnnotation *)pinAtPoint:(NSPoint)point {
+    NSArray *features = [self.mapView visibleFeaturesAtPoint:point];
+    NSString *title;
+    for (id <MGLFeature> feature in features) {
+        if (!title) {
+            title = [feature attributeForKey:@"name_en"] ?: [feature attributeForKey:@"name"];
+        }
+    }
+    
     DroppedPinAnnotation *annotation = [[DroppedPinAnnotation alloc] init];
     annotation.coordinate = [self.mapView convertPoint:point toCoordinateFromView:self.mapView];
-    annotation.title = @"Dropped Pin";
+    annotation.title = title ?: @"Dropped Pin";
     _spellOutNumberFormatter.numberStyle = NSNumberFormatterSpellOutStyle;
     if (_showsToolTipsOnDroppedPins) {
         NSString *formattedNumber = [_spellOutNumberFormatter stringFromNumber:@(++_droppedPinCounter)];
@@ -704,6 +713,10 @@ static const CLLocationCoordinate2D WorldTourDestinations[] = {
         DroppedPinAnnotation *droppedPin = annotation;
         [droppedPin pause];
     }
+}
+
+- (CGFloat)mapView:(MGLMapView *)mapView alphaForShapeAnnotation:(MGLShape *)annotation {
+    return 0.8;
 }
 
 @end